Refactor web module: Extract JavaScript and cleanup server logic
- Migrate inline scripts for table rows and periodic job charts into separate files.
- Organize handlers and server initialization into the `web` package.
- Adjust cron schedule for demonstration job to 15 seconds.
