+++ web/jobs_table_render.go
@@ -1,4 +1,4 @@
-package main
+package web
 
 import (
 	_ "embed"
@@ -12,6 +12,12 @@ import (
 
 //go:embed assets/table_styles.css
 var tableStyles string
+
+//go:embed assets/table_rows.js
+var tableRows string
+
+//go:embed assets/periodic_job_row.js
+var periodicJobRow string
 
 const jobEvent = "job-update"
 
@@ -67,82 +73,7 @@ func renderJobsTable(jobs []jobpro.JobRun, resultCounts map[string]int) string {
 // renderJobsTableRows renders just the table rows - for HTMX updates
 func renderJobsTableRows(b *element.Builder, jobs []jobpro.JobRun, resultCounts map[string]int) (x any) {
 	// Add JavaScript for expand/collapse functionality and load more
-	b.Script().T(`
-	function toggleJobResults(jobId) {
-		const resultRows = document.querySelectorAll('.job-result-row[data-job-id="' + jobId + '"]');
-		const loadMoreRows = document.querySelectorAll('.load-more-row.job-' + jobId);
-		const toggleBtn = document.querySelector('.toggle-btn[data-job-id="' + jobId + '"]');
-		const isExpanded = toggleBtn.classList.contains('expanded');
-		
-		// Toggle result rows
-		resultRows.forEach(row => {
-			row.style.display = isExpanded ? 'none' : '';
-		});
-		
-		// Toggle load more rows
-		loadMoreRows.forEach(row => {
-			row.style.display = isExpanded ? 'none' : '';
-		});
-		
-		toggleBtn.classList.toggle('expanded');
-		toggleBtn.textContent = isExpanded ? '▶' : '▼';
-		
-		// Store state in localStorage
-		const expandedJobs = JSON.parse(localStorage.getItem('expandedJobs') || '{}');
-		expandedJobs[jobId] = !isExpanded;
-		localStorage.setItem('expandedJobs', JSON.stringify(expandedJobs));
-	}
-	
-	// Restore expanded state after HTMX update
-	document.addEventListener('htmx:afterSwap', function() {
-		const expandedJobs = JSON.parse(localStorage.getItem('expandedJobs') || '{}');
-		Object.entries(expandedJobs).forEach(([jobId, isExpanded]) => {
-			if (isExpanded) {
-				const toggleBtn = document.querySelector('.toggle-btn[data-job-id="' + jobId + '"]');
-				if (toggleBtn && !toggleBtn.classList.contains('expanded')) {
-					toggleJobResults(jobId);
-				}
-			}
-		});
-		
-		// Ensure all flex containers are properly styled
-		const allFlexContainers = document.querySelectorAll('td > div[style*="flex"]');
-		allFlexContainers.forEach(container => {
-			// Force re-apply flex styles in case they got lost
-			container.style.display = 'flex';
-			container.style.alignItems = 'center';
-			container.style.gap = '0.5rem';
-		});
-	});
-	
-	function loadMoreResults(button) {
-		const jobId = button.getAttribute('data-job-id');
-		const offset = button.getAttribute('data-offset');
-		const loadMoreRow = button.closest('tr');
-		
-		fetch('/jobs/results/' + jobId + '?offset=' + offset)
-			.then(response => response.text())
-			.then(html => {
-				// Insert the new rows before the load more row
-				loadMoreRow.insertAdjacentHTML('beforebegin', html);
-				// Remove the old load more row
-				loadMoreRow.remove();
-				// Make sure new rows are visible if job is expanded
-				const toggleBtn = document.querySelector('.toggle-btn[data-job-id="' + jobId + '"]');
-				if (toggleBtn && toggleBtn.classList.contains('expanded')) {
-					const newRows = document.querySelectorAll('.job-result-row[data-job-id="' + jobId + '"], .load-more-row.job-' + jobId);
-					newRows.forEach(row => {
-						row.style.display = '';
-					});
-				}
-			})
-			.catch(error => {
-				console.error('Error loading more results:', error);
-				button.textContent = 'Error loading results';
-				button.disabled = true;
-			});
-	}
-	`)
+	b.Script().T(tableRows)
 
 	// Track results per job
 	currentJobID := ""
@@ -212,7 +143,7 @@ func renderJobsTableRows(b *element.Builder, jobs []jobpro.JobRun, resultCounts 
 						b.Div("style", "display: flex; align-items: center; gap: 0.5rem;").R(
 							b.Span("class", "toggle-btn", "data-job-id", job.JobID,
 								"onclick", "toggleJobResults('"+job.JobID+"')",
-								"style", "cursor: pointer; font-size: 0.8rem; user-select: none; flex-shrink: 0;").T("▶"),
+								"style", "cursor: pointer; font-size: 0.8rem; user-select: none; flex-shrink: 0;").T("&#9658;"),
 							b.Span("style", "flex-grow: 1;").T(job.JobName),
 							// Add result count indicator for periodic jobs
 							b.Wrap(func() {
@@ -306,98 +237,8 @@ func renderJobsTableRows(b *element.Builder, jobs []jobpro.JobRun, resultCounts 
 									b.Canvas("id", "chart-"+job.JobID, "style", "display: block; max-height: 60px;").T(""),
 								),
 								// Script to fetch and render chart data
-								b.Script().T(`
-								(function() {
-									const chartId = 'chart-`+job.JobID+`';
-									const canvas = document.getElementById(chartId);
-									if (!canvas) return;
-
-									// Fetch job history
-									fetch('/jobs/history/`+job.JobID+`')
-										.then(response => response.json())
-										.then(data => {
-											if (!data || data.length === 0) {
-												canvas.style.display = 'none';
-												const successRateContainer = document.getElementById('success-rate-`+job.JobID+`');
-												if (successRateContainer) {
-													successRateContainer.innerHTML = '<span style="color: #999;">No runs</span>';
-												}
-												return;
-											}
-											
-											// Calculate success rate
-											const successfulRuns = data.filter(d => d.Status === 'complete').length;
-											const successRate = Math.round((successfulRuns / data.length) * 100);
-											
-											// Display success rate
-											const successRateContainer = document.getElementById('success-rate-`+job.JobID+`');
-											if (successRateContainer) {
-												const rateColor = successRate >= 80 ? '#4ade80' : successRate >= 50 ? '#fbbf24' : '#ef4444';
-												successRateContainer.innerHTML = 
-													'<div style="text-align: center;">' +
-													'<div style="font-size: 1.1em; font-weight: bold; color: ' + rateColor + ';">' + successRate + '%</div>' +
-													'<div style="font-size: 0.8em; color: #666;">success</div>' +
-													'</div>';
-											}
-											
-											// Prepare chart data
-											const labels = data.slice().reverse().map((_, idx) => idx + 1);
-											const durations = data.slice().reverse().map(d => d.Duration / 1000000); // Convert to ms
-											const colors = data.slice().reverse().map(d => 
-												d.Status === 'complete' ? '#4ade80' : '#ef4444'
-											);
-											
-											// Create chart
-											new Chart(canvas, {
-												type: 'line',
-												data: {
-													labels: labels,
-													datasets: [{
-														data: durations,
-														fill: true,
-														backgroundColor: 'rgba(74, 222, 128, 0.2)',
-														borderColor: '#4ade80',
-														borderWidth: 2,
-														pointBackgroundColor: colors,
-														pointBorderColor: colors,
-														pointRadius: 4,
-														pointHoverRadius: 6,
-														tension: 0.3
-													}]
-												},
-												options: {
-													responsive: true,
-													maintainAspectRatio: false,
-													plugins: {
-														legend: { display: false },
-														tooltip: {
-															callbacks: {
-																label: function(context) {
-																	return context.parsed.y.toFixed(1) + ' ms';
-																}
-															}
-														}
-													},
-													scales: {
-														x: { 
-															display: false,
-															grid: { display: false }
-														},
-														y: { 
-															display: false,
-															grid: { display: false },
-															beginAtZero: true
-														}
-													}
-												}
-											});
-										})
-										.catch(error => {
-											console.error('Error fetching job history:', error);
-											canvas.style.display = 'none';
-										});
-								})();
-								`),
+								b.Script().T(`var jobID = `+job.JobID+`;`,
+									periodicJobRow),
 							),
 						)
 					} else {
